job_agent/requirements.txt
flask
sqlalchemy
flask-sqlalchemy
python-dotenv
python-jobspy
google-generativeai
python-docx
requests
markdown
beautifulsoup4
